// packages/shared/src/scoring/reason-codes.d.ts
import type { ReasonCode } from '../types/trust';
export declare const REASON_CODES: {
    OB_CONNECTED: ReasonCode;
    TXN_HISTORY_6M: ReasonCode;
    TXN_HISTORY_3M: ReasonCode;
    NAME_MATCH: ReasonCode;
    DOCUMENT_UPLOADED: ReasonCode;
    ADDRESS_VERIFIED: ReasonCode;
    SELF_DECLARED_ONLY: ReasonCode;
    SALARY_RECURRING: ReasonCode;
    GIG_INCOME_CONSISTENT: ReasonCode;
    LOW_INCOME_VARIANCE: ReasonCode;
    NET_INFLOW_POSITIVE: ReasonCode;
    HIGH_INCOME_VARIANCE: ReasonCode;
    NO_RECURRING_INCOME: ReasonCode;
    RENT_INCOME_HEALTHY: ReasonCode;
    RESIDUAL_BALANCE_HEALTHY: ReasonCode;
    SAVINGS_BUFFER: ReasonCode;
    LOW_OVERDRAFT: ReasonCode;
    NO_RETURNED_PAYMENTS: ReasonCode;
    HIGH_RENT_RATIO: ReasonCode;
    OVERDRAFT_DEPENDENT: ReasonCode;
    RENT_RECURRING: ReasonCode;
    RENT_AMOUNT_CONSISTENT: ReasonCode;
    RENT_TIMING_CONSISTENT: ReasonCode;
    RENT_HISTORY_6M: ReasonCode;
    TENANCY_DOCUMENT: ReasonCode;
    NO_RENT_EVIDENCE: ReasonCode;
    SAVINGS_TREND_UP: ReasonCode;
    INCOME_CONTINUITY: ReasonCode;
    LOW_SPEND_VOLATILITY: ReasonCode;
    DEBT_SERVICING_STABLE: ReasonCode;
    STRONG_CASHFLOW: ReasonCode;
    PROFILE_CONSISTENT: ReasonCode;
    IDENTITY_DOCUMENT: ReasonCode;
    ADDRESS_CONSISTENT: ReasonCode;
    NAME_MISMATCH: ReasonCode;
};
//# sourceMappingURL=reason-codes.d.ts.map